Refactor DI
